// src/types/homeHero.ts
export interface heroTopSmallBanner {
  backgroundColor: string;
  backgroundPhoto: string;
  productPhoto: string;
  title: string;
  price: number;
  buttonText: string;
  link: string;
  _id: string;
}
export interface heroBottomSmallBanner {
  backgroundColor: string;
  backgroundPhoto: string;
  productPhoto: string;
  title: string;
  price: number;
  buttonText: string;
  link: string;
  _id: string;
  offerTag: string;
}